+++ src/commands/sysinfo.js
@@ -86,10 +86,12 @@ const _determineColor = (mem) => {
  *
  * Available arguments:
  * - `-c`: Display computer information.
+ * - `-u`: Display user information.
  * - `-s`: Display system resources.
  * - `-a`: Display advanced information.
  * - `-e`: Display environment variables.
  * - `--all`: Display all system information that is available.
+ * - `--rm-tip`: Don't display the tip.
  *
  * @param  {...string} args Arguments to modify the behavior of `sysinfo`.
  */
@@ -98,6 +100,7 @@ const sysinfo = (...args) => {
     // Initialize arguments
     // Arguments to modify what is shown
     const computerInfo = args?.includes("-c") || args?.includes("/c");
+    const userInfo = args?.includes("-u") || args?.includes("/u");
     const sysResource = args?.includes("-s") || args?.includes("/s");
     const advancedInfo = args?.includes("-a") || args?.includes("/a");
     const envVars = args?.includes("-e") || args?.includes("/e");
@@ -108,7 +111,8 @@ const sysinfo = (...args) => {
     const rmTip = args?.includes("--rm-tip") || args?.includes("/rm-tip");
 
     // In case no arguments were passed to modify what was shown, show that
-    const defaultDisplay = !computerInfo && !sysResource && !advancedInfo && !envVars && !all;
+    const defaultDisplay =
+      !computerInfo && !userInfo && !sysResource && !advancedInfo && !envVars && !all;
 
     // If the user either requested everything, just the computer info, or the default display
     if (all || computerInfo || defaultDisplay) {
@@ -121,6 +125,29 @@ const sysinfo = (...args) => {
       console.log(`Computer name: ${chalk.italic(os.hostname())}`);
       console.log(`Locale: ${chalk.italic(Intl.DateTimeFormat().resolvedOptions().locale)}`);
 
+      console.log();
+    }
+
+    // If the user either requested everything, just the user info, or the default display
+    if (all || userInfo || defaultDisplay) {
+      console.log(`${chalk.bold.underline("User Information")}`);
+
+      // Get all properties from userInfo()
+      const { gid, homedir, shell, uid, username } = os.userInfo();
+
+      // Log basic user information
+      console.log(`Username: ${chalk.bold(username)}`);
+      console.log(`Home directory: ${chalk.bold(homedir)}`);
+      console.log(`Temporary directory: ${chalk.bold(os.tmpdir())}`);
+
+      // If the OS is Windows, GID, shell and UID are -1/null; so this is only shown to operating systems other than Windows
+      if (process.platform !== "win32") {
+        console.log(`GID (group identifier): ${chalk.bold(gid)}`);
+        console.log(`UID (user identifier): ${chalk.bold(uid)}`);
+        console.log(`Shell: ${chalk.bold(shell)}`);
+      }
+
+      // Newline and return
       console.log();
     }